clean up switch

+++ proc/11-cleanup.sh
@@ -1,6 +1,10 @@
 # Clean up!
 
 abinfo "Pre-build clean up..."
-rm -rf abdist
-rm -rf abscripts
-rm -f abspec
+if bool $ABCLEAN; then
+    rm -rf abdist
+    rm -rf abscripts
+    rm -f abspec
+else 
+    true
+done
